cdk/requirements.txt
aws-cdk.core==1.204.0
boto3==1.34.32
botocore==1.34.32